HSEARCH-2260 Moved schema translation code outside of IndexManager
This cleans up the code and lays the groundwork for (later) introducing the schema export tools.
