+++ elasticsearch/src/main/resources/META-INF/services/org.hibernate.search.elasticsearch.schema.impl.ElasticsearchSchemaTranslator
@@ -1,0 +1,1 @@
+org.hibernate.search.elasticsearch.schema.impl.DefaultElasticsearchSchemaTranslator
